Break out skater comparison component
