+++ components/Game/index.ts
@@ -1,9 +1,11 @@
 import DivisionStandings from './Preview/DivisionStandings';
+import SkaterComparison from './Preview/SkaterComparison';
 import TeamsBlock from './Preview/TeamsBlock';
 import TeamStats from './Preview/TeamStats';
 
 export {
   DivisionStandings,
+  SkaterComparison,
   TeamsBlock,
   TeamStats
 };
